+++ CHANGELOG.md
@@ -20,11 +20,17 @@ That said, these are more guidelines rather than hard rules, though the project 
 
 ---
 
+## Unreleased
+
+### Bug Fixes
+
+- [#2131](https://github.com/ClementTsang/bottom/pull/2131): Fix the config not accepting `Memory`, `Memory%`, `Total Read`, and `Total Write` process column aliases that the config schema advertised.
+
 ## 0.14.3 - 2026-07-01
 
 ### Other
 
-- [#2119](https://github.com/ClementTsang/bottom/pull/2119): Workaround change for draw behaviour that caused Kitty with `cursor_trail` to occasionally use more CPU.
+- [#2119](https://github.com/ClementTsang/bottom/pull/2119): Add workaround for draw behaviour that caused Kitty with `cursor_trail` to occasionally use more CPU.
 
 ## 0.14.2 - 2026-06-27